python/requirements.txt
# Core Dependencies
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2

# Web Service (Flask)
flask==3.0.0
flask-cors==4.0.0

# Data Analysis (Opcional - para uso avanzado)
pandas==2.1.4
numpy==1.26.2

# Utilities
requests==2.31.0

